+++ netbox/utilities/tests/test_utils.py
@@ -1,7 +1,8 @@
 from django.http import QueryDict
 from django.test import TestCase
 
-from utilities.utils import deepmerge, dict_to_filter_params, normalize_querydict
+from utilities.data import deepmerge
+from utilities.utils import dict_to_filter_params, normalize_querydict
 
 
 class DictToFilterParamsTest(TestCase):
